Refactor annotation specs diff modules
